价格bug修改

+++ app/src/main/java/com/promote/ebingoo/InformationActivity/ProductInfoActivity.java
@@ -135,7 +135,7 @@ public class ProductInfoActivity extends Activity implements View.OnClickListene
         }
         prdinfocompanytv.setText(infoBean.getCompany_name());
         pi_title_tv.setText(infoBean.getTitle());
-        pi_price_tv.setText((Integer.parseInt(infoBean.getPrice()) == 0) ? getString(R.string.price_zero) : infoBean.getPrice() + "");
+        pi_price_tv.setText((Double.parseDouble(infoBean.getPrice()) == 0) ? getString(R.string.price_zero) : infoBean.getPrice() + "");
         productinfolooknumtv.setText(infoBean.getRead_num() + "");
         if (infoBean.getUnit() != null) {//起售标准
             pi_min_sell_num.setText(infoBean.getMin_sell_num() + "" + infoBean.getUnit());
